requirements.txt
# Основные зависимости
aiogram>=3.22.0
python-dotenv>=1.2.1
aiofiles>=24.1.0
Pillow>=12.0.0
openai==1.57.0
asyncpg>=0.29.0
aiohttp>=3.9.3

# Тестовые зависимости 
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
watchdog>=4.0.0